+++ src/rovr/classes/textual_options.py
@@ -1,11 +1,12 @@
 from os import DirEntry
-from typing import Literal
+from typing import Literal, NamedTuple
 
 from textual.content import Content, ContentText
+from textual.widgets import SelectionList
 from textual.widgets.option_list import Option
 from textual.widgets.selection_list import Selection
 
-from rovr.functions.path import compress
+from rovr.functions.path import normalise
 
 
 class PinnedSidebarOption(Option):
@@ -62,6 +63,7 @@ class FileListSelectionWidget(Selection):
         icon: list[str],
         label: str,
         dir_entry: DirEntry,
+        clipboard: SelectionList,
         disabled: bool = False,
     ) -> None:
         """
@@ -82,16 +84,35 @@ class FileListSelectionWidget(Selection):
 
         # Create prompt by combining cached icon content with label
         prompt = FileListSelectionWidget._icon_content_cache[cache_key] + Content(label)
+        dir_entry_path = normalise(dir_entry.path)
+        if any(
+            dir_entry_path == clipboard_val.path for clipboard_val in clipboard.selected
+        ):
+            prompt = prompt.stylize("dim")
         self.dir_entry = dir_entry
         this_id = str(id(self))
 
         super().__init__(
             prompt=prompt,
+            # this is kinda required for FileList.get_selected_object's select mode
+            # because it gets selected (which is dictionary of values)
+            # which it then queries for `id` (because there's no way to query for
+            # values directly)
             value=str(this_id),
             id=str(this_id),
             disabled=disabled,
         )
+        self._prompt: Content
         self.label = label
+
+    @property
+    def prompt(self) -> Content:
+        return self._prompt
+
+
+class ClipboardSelectionValue(NamedTuple):
+    path: str
+    type_of_selection: Literal["copy", "cut"]
 
 
 class ClipboardSelection(Selection):
@@ -119,11 +140,17 @@ class ClipboardSelection(Selection):
             )
         super().__init__(
             prompt=prompt,
-            value=compress(f"{text}-{type_of_selection}"),
-            id=compress(text),
+            value=ClipboardSelectionValue(text, type_of_selection),
+            # in the future, if we want persistent clipboard,
+            # we will have to switch to use path.compress
+            id=str(id(self)),
         )
         self.initial_prompt = prompt
-        self.path = text
+
+    @property
+    def value(self) -> ClipboardSelectionValue:
+        """The value for this selection."""
+        return self._value
 
 
 class KeybindOption(Option):