+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## new features
 - cli: add `--config-folder` option to specify custom config folder #185
+- filelist: dim files/folders that are cut in clipboard #188
 
 <details>
 <summary><h1>[v0.7.0.dev1]</h1></summary>